+++ src/Color/Factory.php
@@ -9,13 +9,12 @@ class Factory
     public static function init($color)
     {
         $color = str_replace(' ', '', $color);
-
         // Definitive types
         if (preg_match('/^(?P<type>(rgba?|hsla?|hsv|#))/i', $color, $match)) {
-            $class = 'OzdemirBurak\\Iris\\Color\\' . ucfirst(strtolower($match['type'] == '#' ? 'hex' : $match['type']));
+            $class = ucfirst(strtolower($match['type'] === '#' ? 'hex' : $match['type']));
+            $class = 'OzdemirBurak\\Iris\\Color\\' . $class;
             return new $class($color);
         }
-
         // Best guess
         if (preg_match('/^#?[a-f0-9]{3}([a-f0-9]{3})?$/i', $color)) {
             return new Hex($color);
@@ -32,7 +31,6 @@ class Factory
         if (preg_match('/^\d{1,3},\d{1,3}%,\d{1,3}%,[0-9\.]+$/', $color)) {
             return new Hsla($color);
         }
-
         // Cannot determine between hsv and hsl
         throw new AmbiguousColorString("Cannot determine color type of '{$color}'");
     }